refactor: move operation handlers to python module
