--- operations/c8y_Coils/c8y_Coils
@@ -1,4 +1,0 @@
-[exec]
-  topic = "c8y/s/ds"
-  on_message = "xy"
-  command = "/etc/tedge/plugins/modbus/c8y_Coils.py"
