--- operations/c8y_ModbusConfiguration/c8y_ModbusConfiguration
@@ -1,4 +1,0 @@
-[exec]
-  topic = "c8y/s/dc/modbus"
-  on_message = "1"
-  command = "/etc/tedge/plugins/modbus/c8y_ModbusConfiguration.py"
