--- operations/c8y_ModbusDevice/c8y_ModbusDevice
@@ -1,4 +1,0 @@
-[exec]
-  topic = "c8y/s/dc/modbus"
-  on_message = "2"
-  command = "/etc/tedge/plugins/modbus/c8y_ModbusDevice.py"
