+++ tedge_modbus/operations/c8y_modbus_configuration.py
@@ -3,9 +3,10 @@
 """
 import json
 import logging
-import sys
 import toml
 from paho.mqtt.publish import single as mqtt_publish
+
+from .context import Context
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(
@@ -13,14 +14,9 @@ logging.basicConfig(
 )
 logger.info("New c8y_ModbusConfiguration operation")
 
-# TODO: Get broker and port from thin-edge options
-BROKER = "localhost"
-PORT = 1883
-CLIENT_ID = "c8y_ModbusConfiguration-operation-client"
-CONFIG_PATH = "/etc/tedge/plugins/modbus/modbus.toml"
 
-try:
-    arguments = sys.argv[1].split(",")
+def run(arguments, context: Context):
+    """Run c8y_ModbusConfiguration operation handler"""
     if len(arguments) != 4:
         raise ValueError(
             f"Expected 4 arguments in smart rest template. Got {len(arguments)}"
@@ -31,8 +27,9 @@ try:
     logger.debug("transmitRate: %d, pollingRate: %d", transmit_rate, polling_rate)
 
     # Get device configuration
-    logger.info("Read mapping toml from %s", CONFIG_PATH)
-    modbus_config = toml.load(CONFIG_PATH)
+    config_path = context.config_dir / "modbus.toml"
+    logger.info("Read mapping toml from %s", config_path)
+    modbus_config = toml.load(config_path)
     logger.debug("Current configuration: %s", modbus_config)
 
     # Update configuration
@@ -40,10 +37,10 @@ try:
     modbus_config["modbus"]["pollinterval"] = polling_rate
 
     # Save to file
-    logger.info("Saving new configuration to %s", CONFIG_PATH)
-    with open(CONFIG_PATH, "w", encoding="utf8") as f:
+    logger.info("Saving new configuration to %s", config_path)
+    with open(config_path, "w", encoding="utf8") as f:
         toml.dump(modbus_config, f)
-    logger.info("New configuration saved to %s", CONFIG_PATH)
+    logger.info("New configuration saved to %s", config_path)
 
     # Update managedObject
     logger.debug("Updating managedObject with new configuration")
@@ -57,11 +54,7 @@ try:
         payload=json.dumps(config),
         qos=1,
         retain=True,
-        hostname=BROKER,
-        port=PORT,
-        client_id=CLIENT_ID,
+        hostname=context.broker,
+        port=context.port,
+        client_id="c8y_ModbusConfiguration-operation-client",
     )
-
-except Exception as e:
-    logger.error("Error: %s", e)
-    sys.exit(1)
